backend/requirements.txt
fastapi==0.111.0
uvicorn==0.26.0
pydantic==2.6.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.0
pandas==2.1.3
numpy==1.26.3
plotly==5.19.0
python-multipart==0.0.6
requests==2.31.0
httpx==0.27.0